update after making separate databaseAccess.jar file
