+++ build.gradle
@@ -75,7 +75,7 @@ dependencies {
     compile group: 'com.amazonaws', name: 'aws-java-sdk-s3', version: '1.11.327'
     compile group: 'com.amazonaws', name: 'aws-java-sdk-lambda', version: '1.11.327'
     compile group: 'com.cloudant', name: 'cloudant-client', version: '2.17.0'
-    compile group: 'org.mongodb', name: 'mongo-java-driver', version: '3.12.8'
+//    compile group: 'org.mongodb', name: 'mongo-java-driver', version: '3.12.8'
     //compile group: 'ch.qos.logback', name: 'logback-classic', version: '1.2.3'
     testCompile group: 'junit', name: 'junit', version: '3.8.1'
     compile fileTree(dir: 'externalJars', include: ['*.jar'])
