--- database.properties
@@ -1,5 +1,0 @@
-host=10.0.0.62
-port=27017
-database=AFCL
-username=user
-password=pw
