+++ mariaDatabase.properties
@@ -1,0 +1,5 @@
+host=10.0.0.62
+port=3307
+database=afcl
+username=user
+password=pw
