+++ mongoDatabase.properties
@@ -1,0 +1,6 @@
+host=10.0.0.62
+port=27017
+database=AFCL
+collection=logs
+username=user
+password=pw
