+++ src/main/java/at/enactmentengine/serverless/main/Local.java
@@ -1,6 +1,6 @@
 package at.enactmentengine.serverless.main;
 
-import at.uibk.dps.database.MongoDBAccess;
+import at.uibk.dps.mongoLogger.MongoDBAccess;
 import ch.qos.logback.classic.Level;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
